+++ src/twisted/mail/_pop3client.py
