+++ src/twisted/mail/test/test_pop3client.py
@@ -9,7 +9,7 @@ from typing import List
 from unittest import skipIf
 from zope.interface import directlyProvides
 
-import twisted.mail.pop3client
+import twisted.mail._pop3client
 
 from twisted.internet import reactor, defer, error, protocol, interfaces
 from twisted.mail.pop3 import AdvancedPOP3Client as POP3Client
@@ -609,22 +609,22 @@ class POP3ClientModuleStructureTests(TestCase):
 
     def test_all(self):
         """
-        twisted.mail.pop3client.__all__ should be empty because all classes
+        twisted.mail._pop3client.__all__ should be empty because all classes
         should be imported through twisted.mail.pop3.
         """
-        self.assertEqual(twisted.mail.pop3client.__all__, [])
+        self.assertEqual(twisted.mail._pop3client.__all__, [])
 
     def test_import(self):
         """
-        Every public class in twisted.mail.pop3client should be available as a
-        member of twisted.mail.pop3 with the exception of
-        twisted.mail.pop3client.POP3Client which should be available as
+        Every public class in twisted.mail._pop3client should be available as
+        a member of twisted.mail.pop3 with the exception of
+        twisted.mail._pop3client.POP3Client which should be available as
         twisted.mail.pop3.AdvancedClient.
         """
         publicClasses = [
             c[0]
             for c in inspect.getmembers(
-                sys.modules["twisted.mail.pop3client"], inspect.isclass
+                sys.modules["twisted.mail._pop3client"], inspect.isclass
             )
             if not c[0][0] == "_"
         ]
